+++ tests/test_campaign.py
@@ -1,0 +1,31 @@
+"""Tests features of the Campaign object."""
+
+import pytest
+from pytest import param
+
+from .conftest import run_iterations
+
+
+@pytest.mark.parametrize(
+    "target_names",
+    [
+        param(["Target_max"], id="max"),
+        param(["Target_min"], id="min"),
+        param(["Target_max_bounded"], id="max_b"),
+        param(["Target_min_bounded"], id="min_b"),
+        param(["Target_match_bell"], id="match_bell"),
+        param(["Target_match_triangular"], id="match_tri"),
+        param(
+            ["Target_max_bounded", "Target_min_bounded", "Target_match_triangular"],
+            id="desirability",
+        ),
+    ],
+)
+@pytest.mark.parametrize("batch_size", [2], ids=["b2"])
+@pytest.mark.parametrize("n_iterations", [2], ids=["i2"])
+def test_get_surrogate(campaign, n_iterations, batch_size):
+    """Test successful extraction of the surrogate model."""
+    run_iterations(campaign, n_iterations, batch_size)
+
+    model = campaign.get_surrogate()
+    assert model is not None, "Something went wrong during surrogate model extraction."