+++ tests/test_searchspace.py
@@ -6,8 +6,7 @@ import pytest
 
 from baybe.constraints import (
     ContinuousCardinalityConstraint,
-    ContinuousLinearEqualityConstraint,
-    ContinuousLinearInequalityConstraint,
+    ContinuousLinearConstraint,
     DiscreteSumConstraint,
     ThresholdCondition,
 )
@@ -153,22 +152,14 @@ def test_invalid_constraint_parameter_combos():
     with pytest.raises(ValueError):
         SearchSpace.from_product(
             parameters=parameters,
-            constraints=[
-                ContinuousLinearEqualityConstraint(
-                    parameters=["c1", "c2", "d1"],
-                )
-            ],
+            constraints=[ContinuousLinearConstraint(["c1", "c2", "d1"], "=")],
         )
 
     # Attempting continuous constraint over hybrid parameter set
     with pytest.raises(ValueError):
         SearchSpace.from_product(
             parameters=parameters,
-            constraints=[
-                ContinuousLinearInequalityConstraint(
-                    parameters=["c1", "c2", "d1"],
-                )
-            ],
+            constraints=[ContinuousLinearConstraint(["c1", "c2", "d1"], "=")],
         )
 
     # Attempting discrete constraint over hybrid parameter set
@@ -199,11 +190,7 @@ def test_invalid_constraint_parameter_combos():
     with pytest.raises(ValueError):
         SearchSpace.from_product(
             parameters=parameters,
-            constraints=[
-                ContinuousLinearInequalityConstraint(
-                    parameters=["c1", "e7", "d1"],
-                )
-            ],
+            constraints=[ContinuousLinearConstraint(["c1", "e7", "d1"], "=")],
         )
 
     # Attempting constraints over parameter sets containing non-numerical discrete