+++ CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.11.1] - 2024-10-01
+### Added
+- Continuous linear constraints have been consolidated in the new
+  `ContinuousLinearConstraint` class
+
+### Changed
+- `get_surrogate` now also returns the model for transformed single targets or
+  desirability objectives
+
 ### Fixed
 - Unsafe name-based matching of columns in `get_comp_rep_parameter_indices`
+
+### Deprecated
+- `ContinuousLinearEqualityConstraint` and `ContinuousLinearInequalityConstraint`
+  replaced by `ContinuousLinearConstraint` with the corresponding `operator` keyword
 
 ## [0.11.0] - 2024-09-09
 ### Breaking Changes